+++ components/controller/src/identifier_controller/mod.rs
@@ -36,13 +36,11 @@ use keri::{
     },
     oobi::{LocationScheme, Role, Scheme},
     prefix::{BasicPrefix, CesrPrimitive, IdentifierPrefix, IndexedSignature, SelfSigningPrefix},
-    processor::notification::{JustNotification, Notification},
     query::{
         query_event::{QueryArgs, QueryArgsMbx, QueryEvent, QueryRoute, QueryTopics, SignedQuery},
         reply_event::ReplyRoute,
     },
 };
-use teliox::{processor::escrow::MissingIssuerEscrow, tel::Tel};
 
 use super::mailbox_updating::ActionRequired;
 use crate::{error::ControllerError, mailbox_updating::MailboxReminder, Controller};
